+++ sageopt/coniclifts/reformulators.py
@@ -21,8 +21,8 @@ from sageopt.coniclifts.cones import Cone
 
 def separate_cone_constraints(A, b, K, dont_sep=None):
     """
-    Given an affine representation {x : A @ x + b \in K}, construct a direct
-    representation {x : \exists y with  G @ [x,y] + h \in K1, y \in K2},
+    Given an affine representation {x : A @ x + b \\in K}, construct a direct
+    representation {x : \\exists y with  G @ [x,y] + h \\in K1, y \\in K2},
     where ``K1`` consists exclusively of cones types in ``dont_sep``.
 
     Parameters
@@ -45,7 +45,7 @@ def separate_cone_constraints(A, b, K, dont_sep=None):
 
     EX1: Suppose the conic system (A,b,K) can be expressed as {x : B @ x == d, G @ x + h >=0, x >=0 }, where G
     is a non-square matrix. Then by calling this function with dont_sep=None, we obtain the conic system
-       { [x,y] : A1 @ [x, y] == b0 } \cap {[x, y] : x >= 0, y >= 0}
+       { [x,y] : A1 @ [x, y] == b0 } \\cap {[x, y] : x >= 0, y >= 0}
     where A1 = [[B, 0], [G, -I]] and b0 = [d, -h]. This conic system is equivalent to (A,b,K) once we
     project onto the x coordinates.
 
@@ -55,19 +55,19 @@ def separate_cone_constraints(A, b, K, dont_sep=None):
 
     EX2: Suppose the conic system (A,b,K) can be expressed as {x : B @ x == d, L <= x <= U }.
     Then by calling this function with dont_sep=None, we obtain the conic system
-    { [x,y,z] : A1 @ [x,y,z] == b0 } \cap {[x,y,z] : y >= 0, z >= 0 }
+    { [x,y,z] : A1 @ [x,y,z] == b0 } \\cap {[x,y,z] : y >= 0, z >= 0 }
     where A1 = [[B, 0, 0], [I, -I, 0], [-I, 0, -I]] and b0 = [b, L, -U]. This conic system is equivalent
     to (A,b,K) once we project out the [y,z] components.
 
     NOTE: here we ended up introducing slack variables for both inequalities in the variable "x", rather
-    than writing {[x,w] : B @ x == b + B @ L, U - L - x - w == 0} \cap {[x,w] : x >= 0, w >= 0 }. This
+    than writing {[x,w] : B @ x == b + B @ L, U - L - x - w == 0} \\cap {[x,w] : x >= 0, w >= 0 }. This
     Second option is still a valid formulation, but it is not one that this function would generate.
 
 
     EX3: Suppose the conic system (A,b,K) can be expressed as  { x : B @ x == d, G @ x <=_{K_prime} h },
     where every entry of G is nonzero, A is not square, and K_prime does not contain the zero cone.
     Then dont_sep={'0'} yields
-    { [x, y] : A1 @ [x, y] == b0 } \cap { [x, y] : y \in K_{prime} }
+    { [x, y] : A1 @ [x, y] == b0 } \\cap { [x, y] : y \\in K_{prime} }
     for A1 = [[B, 0], [-G, -I]] and b0 = [d, -h].
 
     Notes
@@ -75,7 +75,7 @@ def separate_cone_constraints(A, b, K, dont_sep=None):
     dont_sep=None is equivalent to dont_sep={'0'}.
 
     Common use-case: Passing dont_sep={'0','+'}. This ensures that any constraints on nonlinear cones
-            (i.e. exp, psd, soc, etc...) are stated simply as "x[indices_j] \in K_j", where
+            (i.e. exp, psd, soc, etc...) are stated simply as "x[indices_j] \\in K_j", where
             "indices_{j1}" and "indices_{j2}" are disjoint whenever j1 \neq j2. This is the main
             modeling requirement for MOSEK's standard form.
 
@@ -116,7 +116,7 @@ def separate_cone_constraints(A, b, K, dont_sep=None):
 
 def dualize_problem(c, A, b, Kp):
     """
-    min{ c @ x : A @ x + b in K} == max{ -b @ y : c = A.T @ y, y in K^\dagger }
+    min{ c @ x : A @ x + b in K} == max{ -b @ y : c = A.T @ y, y in K^\\dagger }
 
     Parameters
     ----------