+++ sageopt/tests/test_interop/test_cvxpy.py
@@ -229,5 +229,5 @@ class TestCVXPY(unittest.TestCase):
         pass
 
 if __name__ == '__main__':
-    import nose2
-    nose2.main()
+    import pytest
+    pytest.main()